+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     * [Issue #1726](https://github.com/openziti/ziti/issues/1726) - Fix some sdk hosting logging
     * [Issue #1725](https://github.com/openziti/ziti/issues/1725) - Fix panic in entity event processing
     * [Issue #652](https://github.com/openziti/ziti/issues/652) - CI support for MacOS arm64
+    * [Issue #1663](https://github.com/openziti/ziti/issues/1663) - Create a single command to create a service
 
 # Release 0.32.1